fix(page-layout): #1032 Puck custom component previewを共有化
